+++ src/ui/buttons/content.jsx
@@ -4,310 +4,466 @@
 
 import { node, Fragment, type ChildType } from "@krakenjs/jsx-pragmatic/src";
 import { LANG } from "@paypal/sdk-constants/src";
-import { PayPalLogo, LOGO_COLOR } from "@paypal/sdk-logos/src";
+import {
+  PayPalRebrandLogo,
+  PayPalLogo,
+  LOGO_COLOR,
+} from "@paypal/sdk-logos/src";
 
 import { Text } from "../text";
 
 type ButtonContentMap = {
   [$Values<typeof LANG>]: {|
-    PoweredBy: ({| logoColor: $Values<typeof LOGO_COLOR> |}) => ChildType,
+    PoweredBy: ({|
+      logoColor: $Values<typeof LOGO_COLOR>,
+      shouldApplyRebrandedStyles?: boolean,
+    |}) => ChildType,
   |},
 };
 
 export const buttonContent: ButtonContentMap = {
   ar: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>مدعوم من </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   bg: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>С подкрепата на </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   cs: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Využívá službu </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   da: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Leveret af </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   de: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Abgewickelt durch </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   el: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Με την υποστήριξη του </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   en: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Powered by </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   es: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Desarrollado por </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   et: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Powered by </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   fi: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Palvelun tarjoaa </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   fr: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Optimisé par </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   he: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
-        <PayPalLogo logoColor={logoColor} /> מופעל על-ידי
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}{" "}
+        מופעל על-ידי
       </Fragment>
     ),
   },
   hu: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Üzemeltető: </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   id: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Ditunjang teknologi </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   it: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Con tecnologia </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   ja: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Powered by </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   ko: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>제공: </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   lt: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Sukurta </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   lv: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Darbojas ar </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   ms: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Dikuasakan oleh </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   nl: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Mogelijk gemaakt door </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   no: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Leveres av </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   pl: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Powered by </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   pt: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Powered by </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   ro: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Cu tehnologia </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   ru: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Обработано </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   si: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>බලගන්වන්නේ </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
         <Text> විසිනි</Text>
       </Fragment>
     ),
   },
   sk: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Používa technológiu </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   sl: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Powered by </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   sq: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Mundësuar nga </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   sv: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Tillhandahålls av </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   th: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>ให้บริการโดย </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   tl: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Pinapagana ng </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   tr: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Çalıştıran </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   vi: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>Được hỗ trợ bởi </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   zh: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>技术支持提供方： </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },
   zh_Hant: {
-    PoweredBy: ({ logoColor }) => (
+    PoweredBy: ({ logoColor, shouldApplyRebrandedStyles = false }) => (
       <Fragment>
         <Text>技術支持： </Text>
-        <PayPalLogo logoColor={logoColor} />
+        {shouldApplyRebrandedStyles ? (
+          <PayPalRebrandLogo logoColor={logoColor} />
+        ) : (
+          <PayPalLogo logoColor={logoColor} />
+        )}
       </Fragment>
     ),
   },